+++ src/infrastructure/index.ts
@@ -1,5 +1,5 @@
-import { OnMemoryTaskRepository } from './onMemoryTaskRepository';
+import { OnMemoryTaskRepository } from './onMemory/onMemoryTaskRepository';
 
 export let onMemoryTaskRepository: OnMemoryTaskRepository = OnMemoryTaskRepository.create();
 
-export * from './onMemoryTaskRepository';
+export * from './onMemory/onMemoryTaskRepository';
